feat: updated backend wrapper
